dts/hikey: add spidev node with DMA enabled
As reported in the binding of PL022 spi controller, each spi slave
that needs DMA must explicitly enable the DMA using the following
property in the DT node of the slave:
	pl022,com-mode = <2>;

In hi6220, the RX FIFO of PL022 does not work properly together with
the k3-dma engine using the default settings; the last received data
remains in the RX FIFO, thus the RX transfer does not complete, the
DMA callback is never triggered and the spi communication hangs.
The RX FIFO cannot be disabled and its watermark level cannot be
changed.
To fix the issue, the driver of PL022 must inform the DMA driver
about the watermark level of the RX FIFO (4 samples). This is done
by adding the following property in the DT node of the spi slave
that uses DMA:
	pl022,rx-level-trig = <1>;

Add a spidev DT node (using DMA) beside the existing one (not using
DMA). They cannot be enabled at the same time.
Add a comment on top of the new DT node about the two properties.

Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>